power: remove unneeded logging code
for: "Failed to acquire lock"

This error became really common as most ROMs are allowing to use
the PowerHAL without the proper information. Apart from that fact
it means nothing as everything works properly anyways, doing
conditional checks for supressed logging is nonsense and thus
should be removed.

The program will now silently handle the error.

Change-Id: I2885dbe133b34b4d4b71aea403069eefa73602f5